update test packages
