+++ src/test/java/uk/gov/hmcts/reform/unspec/handler/callback/camunda/docmosis/GenerateAcknowledgementOfServiceCallbackHandlerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.unspec.handler.callback;
+package uk.gov.hmcts.reform.unspec.handler.callback.camunda.docmosis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -11,13 +11,13 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.unspec.callback.CallbackParams;
-import uk.gov.hmcts.reform.unspec.handler.callback.camunda.docmosis.GenerateDirectionsQuestionnaireCallbackHandler;
+import uk.gov.hmcts.reform.unspec.handler.callback.BaseCallbackHandlerTest;
 import uk.gov.hmcts.reform.unspec.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.unspec.model.CaseData;
 import uk.gov.hmcts.reform.unspec.model.documents.CaseDocument;
 import uk.gov.hmcts.reform.unspec.model.documents.Document;
 import uk.gov.hmcts.reform.unspec.sampledata.CaseDataBuilder;
-import uk.gov.hmcts.reform.unspec.service.docmosis.dq.DirectionsQuestionnaireGenerator;
+import uk.gov.hmcts.reform.unspec.service.docmosis.aos.AcknowledgementOfServiceGenerator;
 
 import java.time.LocalDateTime;
 
@@ -27,23 +27,23 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.unspec.callback.CallbackType.ABOUT_TO_SUBMIT;
-import static uk.gov.hmcts.reform.unspec.model.documents.DocumentType.DIRECTIONS_QUESTIONNAIRE;
+import static uk.gov.hmcts.reform.unspec.model.documents.DocumentType.ACKNOWLEDGEMENT_OF_SERVICE;
 import static uk.gov.hmcts.reform.unspec.model.documents.DocumentType.SEALED_CLAIM;
 import static uk.gov.hmcts.reform.unspec.utils.ElementUtils.wrapElements;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = {
-    GenerateDirectionsQuestionnaireCallbackHandler.class,
+    GenerateAcknowledgementOfServiceCallbackHandler.class,
     JacksonAutoConfiguration.class,
     CaseDetailsConverter.class
 })
-class GenerateDirectionsQuestionnaireCallbackHandlerTest extends BaseCallbackHandlerTest {
+class GenerateAcknowledgementOfServiceCallbackHandlerTest extends BaseCallbackHandlerTest {
 
     public static final CaseDocument DOCUMENT = CaseDocument.builder()
         .createdBy("John")
         .documentName("document name")
         .documentSize(0L)
-        .documentType(DIRECTIONS_QUESTIONNAIRE)
+        .documentType(ACKNOWLEDGEMENT_OF_SERVICE)
         .createdDatetime(LocalDateTime.now())
         .documentLink(Document.builder()
             .documentUrl("fake-url")
@@ -53,29 +53,29 @@ class GenerateDirectionsQuestionnaireCallbackHandlerTest extends BaseCallbackHan
         .build();
 
     @MockBean
-    private DirectionsQuestionnaireGenerator directionsQuestionnaireGenerator;
+    private AcknowledgementOfServiceGenerator acknowledgementOfServiceGenerator;
 
     @Autowired
-    private GenerateDirectionsQuestionnaireCallbackHandler handler;
+    private GenerateAcknowledgementOfServiceCallbackHandler handler;
 
     @Autowired
     private final ObjectMapper mapper = new ObjectMapper();
 
     @BeforeEach
     void setup() {
-        when(directionsQuestionnaireGenerator.generate(any(CaseData.class), anyString())).thenReturn(DOCUMENT);
+        when(acknowledgementOfServiceGenerator.generate(any(CaseData.class), anyString())).thenReturn(DOCUMENT);
     }
 
     @Test
     void shouldAddDocumentToSystemGeneratedDocuments_whenInvoked() {
-        CaseData caseData = CaseDataBuilder.builder().atStateRespondedToClaim()
+        CaseData caseData = CaseDataBuilder.builder().atStateServiceConfirmed()
             .systemGeneratedCaseDocuments(wrapElements(CaseDocument.builder().documentType(SEALED_CLAIM).build()))
             .build();
         CallbackParams params = callbackParamsOf(caseData, ABOUT_TO_SUBMIT);
 
         var response = (AboutToStartOrSubmitCallbackResponse) handler.handle(params);
 
-        verify(directionsQuestionnaireGenerator).generate(caseData, "BEARER_TOKEN");
+        verify(acknowledgementOfServiceGenerator).generate(caseData, "BEARER_TOKEN");
 
         CaseData updatedData = mapper.convertValue(response.getData(), CaseData.class);
