+++ src/test/java/uk/gov/hmcts/reform/unspec/handler/callback/tasks/EndBusinessProcessTaskHandlerTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.unspec.service.tasks.handler;
+package uk.gov.hmcts.reform.unspec.handler.callback.tasks;
 
 import org.camunda.bpm.client.task.ExternalTask;
 import org.camunda.bpm.client.task.ExternalTaskService;
